Settings: lockscreen wallpaper crash fix
If a user backs out fast enough after setting a
custom wallpaper, the fragment becomes detached
from the activity, so any calls to getActivity()
were throwing exceptions.

The solution is to use the application Context
for toassts to still inform the user of the
status, and not update any UI elements
after the fragment is no longer attached.

Change-Id: Ia9f61184fb484d3af25565bc3c56e8b7b63ff4eb
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	src/com/android/settings/cyanogenmod/LockscreenInterface.java